Fixed mana gui and implemented stamina

+++ FantasyFighter/Assets/Scripts/Unit/UnitScript.cs
@@ -17,7 +17,7 @@ public class UnitScript : MonoBehaviour {
     public int maxHP;
     public int currentHP;
     public int mana;
-    public int stamina;
+    public float stamina;
 
     private Animator anim;
     private Rigidbody unitRigidbody;
@@ -33,7 +33,7 @@ public class UnitScript : MonoBehaviour {
         currentHP = 100;
         maxHP = 100;
         mana = 100;
-        stamina = 100;
+        stamina = 100f;
 
         floorMask = LayerMask.GetMask("Floor");
         anim = GetComponent<Animator>();
